% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cimadata-package.R
\docType{package}
\name{cimadata-package}
\alias{cimadata}
\alias{cimadata-package}
\title{cimadata: Data utility for the Center of Atmospheric and Oceanic Research}
\description{
Get access to data in the CIMA servers.
}
\author{
\strong{Maintainer}: Elio Campitelli \email{elio.campitelli@cima.fcen.uba.ar} (\href{https://orcid.org/0000-0002-7742-9230}{ORCID})

}
\keyword{internal}
