% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmip6.R
\name{cmip_available}
\alias{cmip_available}
\title{Lists available data}
\usage{
cmip_available(base_dir = cmip_folder_get())
}
\arguments{
\item{base_dir}{directory of the CMIP6 databse}
}
\description{
Lists available data
}
