% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmip6.R
\name{cmip_consolidate}
\alias{cmip_consolidate}
\title{Concatenates NetCDF files belonging to different members of the same experiment.}
\usage{
cmip_consolidate(files = NULL, base_dir)
}
\arguments{
\item{files}{character vector of filenames}

\item{base_dir}{(optional) base directory of the CMIP6 file structure to consolidate every file}
}
\description{
Concatenates NetCDF files belonging to different members of the same experiment.
}
