% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/passwords.R
\name{cmip_default_user}
\alias{cmip_default_user}
\alias{cmip_default_user_get}
\alias{cmip_default_user_set}
\title{Default users}
\usage{
cmip_default_user_set(user = NULL)

cmip_default_user_get()
}
\arguments{
\item{user}{user to use as the default (`NULL` means no default user).}
}
\value{
`cmip_default_user_get()` returns the default user set with `cmip_default_user_get()`
or the user if there's only one user saved with `cmip_key_set()`.

`cmip_default_user_get()` sets the default user (for the current session)
and returns the supplied user invisibly.
}
\description{
Default users
}
\section{Functions}{
\itemize{
\item \code{cmip_default_user_set}: Sets the default user.

\item \code{cmip_default_user_get}: Gets the default user.
}}

