% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmip6.R
\name{cmip_folder_set}
\alias{cmip_folder_set}
\alias{cmip_folder_get}
\title{Define y obtiene la carpeta raiz de los datos CMIP6}
\usage{
cmip_folder_set(base_dir)

cmip_folder_get()
}
\arguments{
\item{base_dir}{carpeta raiz de los datos de CMIP6}
}
\description{
Define y obtiene la carpeta raiz de los datos CMIP6
}
