% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/passwords.R
\name{cmip_key}
\alias{cmip_key}
\alias{cmip_key_remove}
\alias{cmip_key_get}
\alias{cmip_key_set}
\title{Password management}
\usage{
cmip_key_set(user = NULL, password = NULL)

cmip_key_get(user)
}
\arguments{
\item{user, password}{username and password. If `NULL`, they will be asked
interactively (recommended).}
}
\description{
Functions to save, get and remove cmip6tinr passwords.
}
\details{
#'
The default `NULL` value for user and password is the recommended method for
secuirity, as otherwise your credentials will be saved in plain text at the
command history.
}
\section{Functions}{
\itemize{
\item \code{cmip_key_set}: Set a new user/password combination

\item \code{cmip_key_get}: Get password for an user.
}}

