% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmip6.R
\name{cmip_mount}
\alias{cmip_mount}
\alias{cmip_unmount}
\title{Mounts and unmounts the CMIP6 database on a local folder}
\usage{
cmip_mount(user, base_dir = cmip_folder_get())

cmip_unmount(base_dir = cmip_folder_get())
}
\arguments{
\item{user}{user}

\item{base_dir}{directory where to mount the folder}
}
\description{
Mounts and unmounts the CMIP6 database on a local folder
}
