% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmip6.R
\name{cmip_size}
\alias{cmip_size}
\title{Gets the total size of the results in megabites}
\usage{
cmip_size(results)
}
\arguments{
\item{results}{a list of results as returned by [cmip_search]}
}
\description{
Gets the total size of the results in megabites
}
