% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cimadata-package.R
\docType{package}
\name{cimadata-package}
\alias{cimadata}
\alias{cimadata-package}
\title{cimadata: Paquete de R para datos del CIMA}
\description{
Utilidades para buscar, descargar datos y para luego leerlos.
}
\author{
\strong{Maintainer}: Elio Campitelli \email{elio.campitelli@cima.fcen.uba.ar} (\href{https://orcid.org/0000-0002-7742-9230}{ORCID})

}
\keyword{internal}
