% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmip6.R
\name{cmip_folder_set}
\alias{cmip_folder_set}
\alias{cmip_folder_get}
\title{Sets and gets CMIP6 folder}
\usage{
cmip_folder_set(base_dir)

cmip_folder_get()
}
\arguments{
\item{base_dir}{directory of the CMIP6 databse}
}
\description{
Sets and gets CMIP6 folder
}
