% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmip6.R
\name{cmip_mount}
\alias{cmip_mount}
\title{Mounts the CMIP6 database on a local folder}
\usage{
cmip_mount(user, base_dir = "~/DATOS/CMIP6")
}
\arguments{
\item{user}{user}

\item{base_dir}{directory where to mount the folder}
}
\description{
Mounts the CMIP6 database on a local folder
}
